Multipath Proposal
We currently have several path-like things in multiformats, ipfs, and libp2p.
This proposal unifies these formats under a single multipath umbrella. This:

1. Gives us a clear, conflict free, central registry for all path namespaces. This
already existed in the form of the multicodec table, this spec just formalizes
the fact that multicodec names map to path components.
2. Makes multiaddr's compact form available to our other path-like formats.
3. Provides a clean way to disambiguate between CIDs, multiaddrs, IPFS paths,
etc. With this spec, a record (e.g. ENS or IPNS) can unambiguously support
arbitrary paths.

---

NOT FOR MERGING

This PR exists for review/discussion. If we decide to go in this direction, I
will create a new multipath repo and PR this spec there.